Refactor Group into multiple files
